Make CSS for mobile cleaner
Removed timestamps in condensed-width view of the directory pages and
made the transition from side-nav to top-nav cleaner and more contained.

Making the navbar responsive was not in the @media tag as I suspected,
but was actually in a flex parameter that did not respond at the same
time or in the same way as the @media tag. At the transition point,
undesired changes occurred and this was resolved by centralizing
everything into the @media tag.