Fix title and update font
The title had a trailing hyphen on pages that weren't posts, so that was
removed from the head.html file.

I didn't love the sans-serif default font, so I added "Open Sans" as the
first font choice. I believe this increases readability and makes it a
much more pleasant site to look at.